Serializer 변경으로 인한 캐스팅 오류 수정 / Profile 데이터 저장 변수 변경

+++ Script/Game/Event/C_StartMatching.cs
@@ -8,7 +8,7 @@ namespace SimulFactory.Game.Event
     {
         public static void StartMatchingC(PcInstance pc, Dictionary<byte,object> param)
         {
-            switch((Define.MATCH_TYPE)(long)param[0])
+            switch((Define.MATCH_TYPE)(int)param[0])
             {
                 case Define.MATCH_TYPE.Normal:
                     NormalMatchSystem.GetInstance().AddPcInsatnce(pc);
